feat: 게시글 조회 기능 구현 (GET: /posts/:id, GET: /posts)
feat: 게시글 조회 기능 구현 (GET: /posts)
- 게시글 id와 게시글 제목만 노출
- 최신 글 순서로 정렬
- 페이지네이션(추가 로드는 20 개 단위)
  - page, pageSize(20) 설정으로 프론트 개발자가 스크롤에 따른 요청을 보낼 때마다 ++page되며 20개씩 게시글 리스트 반환
  - middleware를 통해서 값이 없거나 유효하지 않은 경우 page 1로 default 값 설정 후 service 로직 호출

feat: 게시글 조회 기능 구현 (GET: /posts/:id)
- 패스워드를 제외한 게시글 정보(id, 제목, 내용, 생성일자, 수정일자) 모두 노출

+++ src/posts/posts.service.ts
@@ -10,6 +10,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { PostEntity } from './entities/post.entity';
 import { Repository } from 'typeorm';
 import { BCRYPT_SALT } from '../environments';
+import { EntityNotFoundException } from '../common/exceptions/exception';
 
 @Injectable()
 export class PostsService {
@@ -26,12 +27,34 @@ export class PostsService {
     return await this.postsRepository.save(createPostDto);
   }
 
-  findAll() {
-    return `This action returns all posts`;
+  async findAll(page = 1) {
+    const take = 20;
+
+    return await this.postsRepository
+      .createQueryBuilder('posts')
+      .select('posts.id')
+      .addSelect('posts.title')
+      .take(take)
+      .skip(take * (page - 1))
+      .getMany();
   }
 
-  findOne(id: number) {
-    return `This action returns a #${id} post`;
+  async findOne(id: number) {
+    const post = await this.postsRepository
+      .createQueryBuilder('posts')
+      .select('posts.id')
+      .addSelect('posts.title')
+      .addSelect('posts.content')
+      .addSelect('posts.created_at')
+      .addSelect('posts.updated_at')
+      .where('posts.id = :id', { id: id })
+      .getOne();
+
+    if (!post) {
+      throw new EntityNotFoundException();
+    }
+
+    return post;
   }
 
   update(id: number, updatePostDto: UpdatePostDto) {
@@ -49,10 +72,11 @@ export class PostsService {
   ): Promise<boolean | undefined> {
     const post = await this.postsRepository.findOneBy({ id: id });
 
-    await Promise.all([
-      this.validatePost(id, post),
-      this.validatePassword(enteredPassword, post.password),
-    ]);
+    if (!post) {
+      throw new EntityNotFoundException();
+    }
+
+    await Promise.all([this.validatePassword(enteredPassword, post.password)]);
 
     return (await this.postsRepository.delete({ id: id })) && true;
   }
@@ -71,13 +95,5 @@ export class PostsService {
         "you don't have permission to access this resource",
       );
     }
-  }
-
-  private validatePost(id: number, post: PostEntity) {
-    if (!post) {
-      throw new NotFoundException();
-    }
-
-    return true;
   }
 }
